refactor(Nav/Conversation): reorganize imports and fix import paths
