+++ client/src/components/Conversations/Conversations.tsx
@@ -1,11 +1,18 @@
 import Conversation from './Conversation';
+import { TConversation } from '@librechat/data-provider';
 
-export default function Conversations({ conversations, moveToTop }) {
+export default function Conversations({
+  conversations,
+  moveToTop,
+}: {
+  conversations: TConversation[];
+  moveToTop: () => void;
+}) {
   return (
     <>
       {conversations &&
         conversations.length > 0 &&
-        conversations.map((convo) => {
+        conversations.map((convo: TConversation) => {
           return (
             <Conversation key={convo.conversationId} conversation={convo} retainView={moveToTop} />
           );
