+++ client/src/components/Conversations/index.ts
@@ -1,0 +1,5 @@
+export { default as Pages } from './Pages';
+export { default as Conversation } from './Conversation';
+export { default as DeleteButton } from './DeleteButton';
+export { default as RenameButton } from './RenameButton';
+export { default as Conversations } from './Conversations';
