+++ client/src/components/Nav/Nav.jsx
@@ -1,16 +1,13 @@
+import { NewChat, NavLinks } from './';
+import { Panel, Spinner } from '~/components';
+import { Conversations, Pages } from '../Conversations';
 import { useCallback, useEffect, useRef, useState } from 'react';
 import { useGetConversationsQuery, useSearchQuery } from '@librechat/data-provider';
 import { useRecoilValue, useSetRecoilState } from 'recoil';
-
-import Conversations from '../Conversations';
-import NavLinks from './NavLinks';
-import NewChat from './NewChat';
-import Pages from '../Conversations/Pages';
-import { Panel, Spinner } from '~/components';
-import { cn } from '~/utils/';
 import { useAuthContext, useDebounce } from '~/hooks';
-import store from '~/store';
 import { localize } from '~/localization/Translation';
+import { cn } from '~/utils/';
+import store from '~/store';
 
 export default function Nav({ navVisible, setNavVisible }) {
   const [isHovering, setIsHovering] = useState(false);
