+++ client/src/routes/Root.jsx
@@ -7,12 +7,11 @@ import {
 } from '@librechat/data-provider';
 
 import MessageHandler from '../components/MessageHandler';
-import MobileNav from '../components/Nav/MobileNav';
-import Nav from '../components/Nav';
+import { Nav, MobileNav } from '../components/Nav';
 import { Outlet } from 'react-router-dom';
-import store from '~/store';
 import { useAuthContext } from '~/hooks/AuthContext';
 import { useSetRecoilState } from 'recoil';
+import store from '~/store';
 
 export default function Root() {
   const [navVisible, setNavVisible] = useState(() => {
